requirements.txt
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
